+++ packages/components/segmented-control/src/segmented-control.tsx
@@ -9,6 +9,7 @@ import type {
 import {
   ui,
   forwardRef,
+  FC,
   useComponentMultiStyle,
   omitThemeProps,
 } from "@yamada-ui/core"
@@ -31,7 +32,6 @@ import {
 import type {
   ChangeEvent,
   ChangeEventHandler,
-  FC,
   FocusEventHandler,
   ReactElement,
 } from "react"
@@ -140,9 +140,10 @@ export const SegmentedControl = forwardRef<SegmentedControlProps, "div">(
       onChange: onChangeProp,
       ...rest
     } = omitThemeProps(mergedProps)
+    const uuid = useId()
 
-    id ??= useId()
-    name ??= `segmented-control-${useId()}`
+    id ??= uuid
+    name ??= `segmented-control-${uuid}`
 
     const onChangeRef = useCallbackRef(onChangeProp)
 
@@ -374,6 +375,9 @@ export const SegmentedControl = forwardRef<SegmentedControlProps, "div">(
   },
 )
 
+SegmentedControl.displayName = "SegmentedControl"
+SegmentedControl.__ui__ = "SegmentedControl"
+
 interface SegmentedControlButtonOptions {
   /**
    * The value of the segmented control button.
@@ -462,6 +466,7 @@ export const SegmentedControlButton = forwardRef<
 )
 
 SegmentedControlButton.displayName = "SegmentedControlButton"
+SegmentedControlButton.__ui__ = "SegmentedControlButton"
 
 interface SegmentedControlCursorProps extends MotionProps {
   className?: string
@@ -497,3 +502,6 @@ const SegmentedControlCursor: FC<SegmentedControlCursorProps> = ({
     />
   )
 }
+
+SegmentedControlCursor.displayName = "SegmentedControlCursor"
+SegmentedControlCursor.__ui__ = "SegmentedControlCursor"