+++ packages/components/table/src/tbody.tsx
@@ -1,3 +1,4 @@
+import { FC } from "@yamada-ui/core"
 import type { TableBodyProps as NativeTableBodyProps } from "@yamada-ui/native-table"
 import { Tbody as NativeTbody, Tr, Td } from "@yamada-ui/native-table"
 import { dataAttr, ariaAttr, handlerAll, runIfFunc } from "@yamada-ui/utils"
@@ -5,7 +6,7 @@ import { useTableContext, render } from "./use-table"
 
 export interface TableBodyProps extends NativeTableBodyProps {}
 
-export const Tbody = ({ ...rest }: TableBodyProps) => {
+export const Tbody: FC = ({ ...rest }: TableBodyProps) => {
   const {
     rows,
     rowProps,
@@ -80,3 +81,6 @@ export const Tbody = ({ ...rest }: TableBodyProps) => {
     </NativeTbody>
   )
 }
+
+Tbody.displayName = "Tbody"
+Tbody.__ui__ = "Tbody"